requirements.txt
a2a-sdk>=0.1.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
azure-servicebus>=7.11.0
azure-cosmos>=4.5.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
pandas>=2.0.0
pyyaml>=6.0.0
httpx>=0.28.1

